made the code tidy
