+++ src/pages/movie/post/index.tsx
@@ -1,9 +1,8 @@
 import React, { FC } from 'react';
 import { PostMovieTemplate } from 'components/templates/postMovieTemplate';
-import { VerificationResponse } from 'types/authResponseTypes';
 import { AuthCheckWrapper } from 'components/organisms';
 
-const PostMovie: FC<VerificationResponse> = () => {
+const PostMovie: FC = () => {
   return (
     <AuthCheckWrapper>
       <PostMovieTemplate />
