+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './inputTypes';
 export * from './eventTypes';
 export * from './youtubeTypes';
-export * from './authResponseTypes';
